JSRPC: Allow RPC without `js_rpc` flag if extending new classes.
We will automatically enable JS RPC on the server side if the entrypoint class extends `DurableObject` or `StatelessService`. We'll also allow it for bare object entrypoints, because this seems harmless.

Historical Durable Object classes which did not extend `DurableObject` will not accept RPC, for fear of exposing methods that weren't intended to be exposed.

The `js_rpc` compat flag can still be used to enable RPC on all Durable Object classes, in order to avoid breaking anyone who is already playing with this functionality, but the flag may eventually go away.
