+++ src/workerd/api/tests/js-rpc-flag.wd-test
@@ -9,10 +9,10 @@ const config :Workerd.Config = (
 
 const mainWorker :Workerd.Worker = (
   compatibilityDate = "2022-09-16",
-  compatibilityFlags = ["experimental", "nodejs_compat"],
+  compatibilityFlags = ["experimental", "nodejs_compat", "js_rpc"],
 
   modules = [
-    (name = "worker", esModule = embed "js-rpc-disabled.js"),
+    (name = "worker", esModule = embed "js-rpc-flag.js"),
   ],
 
   durableObjectNamespaces = [
